🚚 Move ShowImage function
