+++ src/Draw/draw.cpp
@@ -14,28 +14,7 @@
 #include "Primitive/I_Primitive.hpp"
 #include "3dDatas/Point3D.hpp"
 #include "Consts/const.hpp"
-
-static void showImage(sf::RenderWindow &window, sf::Image &image) {
-    sf::Sprite sp;
-    sf::Texture txt;
-
-    txt.loadFromImage(image);
-    sp.setTexture(txt);
-    window.clear();
-    window.draw(sp);
-    window.display();
-}
-
-void displayImage(sf::RenderWindow &window, sf::Image &image) {
-    while (window.isOpen()) {
-        sf::Event event;
-        while (window.pollEvent(event)) {
-            if (event.type == sf::Event::Closed)
-                window.close();
-        }
-        showImage(window, image);
-    }
-}
+#include "Generation/tools.hpp"
 
 static double getLuminescence(RayTracer::Point3D &intersection,
     std::unique_ptr<Light> &Light) {
